update renamed config
