+++ lib/config-loader.js
@@ -14,7 +14,7 @@ const path = require('path')
 const yaml = require('js-yaml')
 const fs = require('fs')
 const utils = require('./utils')
-const aioConfig = require('@adobe/aio-cli-config')
+const aioConfig = require('@adobe/aio-lib-core-config')
 
 const debug = require('debug')('cna-scripts:config-loader')
