+++ test/__mocks__/@adobe/aio-lib-core-config.js
