+++ test/scripts/build.actions.test.js
@@ -20,7 +20,7 @@ utils.installDeps = jest.fn()
 // we are mocking zipfolder because of mock-fs not working properly
 // with streams, this might change in future versions of mock-fs
 utils.zipFolder = jest.fn((dir, out) => fs.writeFile(out, 'mock content'))
-const mockAIOConfig = require('@adobe/aio-cli-config')
+const mockAIOConfig = require('@adobe/aio-lib-core-config')
 
 let scripts
 let buildDir
